move secret_manager file
